rag_assistant/requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
torch
faiss-cpu
mlflow
python-dotenv
multipart